Optimize slave cleanup
Today we make calls to Azure while cleaning up the slaves.  This is problematic because calling Jenkins removeNode will (for some reason) call check on the node retention strategy for other nodes.  While removeNode is running, the Queue is locked.  If upon the nested check, we decide to remove a node, the call to Azure, which might take a number of seconds, will block the Queue.  This means in the meantime no builds can start, etc.  This is especially problematic becuase if other triggers/hooks start running, they may also block.  We may end up with a full set of HTTP requests being handled and waiting, meaning the website will appear down while the nodes are being deleted.

This change cleans this up.  First, I removed setDeleteSlave and replaced it with a separate CleanupAction enumeration to avoid overloading and hacky behavior.  Then, along the retention strategy path I removed all calls to delete the slave from azure, instead opting to set the machine offline and wait for the clean up task to come along and remove the node, which is done in a separate thread that will not end up with a blocking call with the queue locked.

Conflicts:
	src/main/java/com/microsoftopentechnologies/azure/AzureCloud.java
	src/main/java/com/microsoftopentechnologies/azure/AzureCloudRetensionStrategy.java
	src/main/java/com/microsoftopentechnologies/azure/AzureManagementServiceDelegate.java
	src/main/java/com/microsoftopentechnologies/azure/AzureSlave.java
	src/main/java/com/microsoftopentechnologies/azure/AzureSlaveCleanUpTask.java